+++ test/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_05_28_041652) do
+ActiveRecord::Schema[8.1].define(version: 2026_07_29_183000) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -37,6 +37,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_05_28_041652) do
   end
 
   create_table "hitch_clients", force: :cascade do |t|
+    t.string "application_type"
     t.string "client_id", null: false
     t.string "client_name", null: false
     t.datetime "created_at", null: false